[2024.04.01] 작업 수정

+++ backend/src/main/java/com/backend/entity/Board.java
@@ -37,7 +37,7 @@ public class Board extends BaseEntity {
 	@JoinColumn(name = "cno")
 	private Category category;
 
-	private int viewCount;
+	private long viewCount;
 
 	@OneToMany(mappedBy = "board", cascade = {
 		CascadeType.ALL}, fetch = FetchType.LAZY, orphanRemoval = true)
@@ -63,7 +63,7 @@ public class Board extends BaseEntity {
 	 *
 	 * @param viewCount
 	 */
-	public void updateViewCount(int viewCount) {
+	public void updateViewCount(long viewCount) {
 		this.viewCount = viewCount;
 	}
